added unified camera module to combine both cam and qr scanner
