+++ app_test.py
@@ -1,5 +1,5 @@
 from flask import Flask, render_template, request, Response, jsonify, redirect, url_for, session
-from camera import UnifiedCamera  # Import the new unified camera
+from unifiedCamera import UnifiedCamera  # Import the new unified camera
 import time
 import os
 from datetime import datetime
