+++ camera.py
@@ -1,248 +1,118 @@
 """
-Unified camera module with built-in QR scanning capability
-@author Salman H.
+This file is for handling the live feed from the camera
+@author
+Salman H.
 """
 import cv2
 import time
 import threading
 import numpy as np
-from pyzbar.pyzbar import decode
-from collections import deque
-import queue
-from picamera2 import Picamera2
+import subprocess
+import tempfile
+import os
+# from picamera2 import Picamera2
 
-class UnifiedCamera:
+class Camera:
     def __init__(self, width=640, height=480, jpeg_quality=70):
-        """Initialize camera with integrated QR scanning"""
-        print(f"Initializing unified camera with size {width}x{height}...")
+        """Initialize camera using libcamera via subprocess with specified quality
 
+        Args:
+            width: Width to resize frames to (default: 320)
+            height: Height to resize frames to (default: 240)
+            jpeg_quality: JPEG compression quality (0-100, default: 70)
+        """
+        print(f"Initializing camera with size {width}x{height}, quality {jpeg_quality}...")
+
+        # Store quality settings
         self.width = width
         self.height = height
         self.jpeg_quality = jpeg_quality
 
-        # Frame management
+        # Create a lock for thread-safe access to camera
         self.frame_lock = threading.Lock()
         self.current_frame = None
-        self.qr_frame = None
+        self.temp_dir = tempfile.mkdtemp()
+        self.temp_file = os.path.join(self.temp_dir, "temp_frame.jpg")
 
-        # QR scanning state
-        self.qr_data_queue = queue.Queue(maxsize=5)  # Store recent QR detections
-        self.last_qr_data = None
-        self.last_qr_time = 0
-        self.qr_cooldown = 10.0  # 10 second cooldown
-        self.qr_scanning_enabled = True
-
-        # Initialize camera
-        self.picam2 = Picamera2()
-        config = self.picam2.create_video_configuration(
-            main={"size": (width * 2, height * 2)},  # High res for display
-            lores={"size": (640, 480), "format": "YUV420"},  # Optimized for QR
-            controls={
-                "FrameRate": 30,
-                "ExposureTime": 8000,
-                "AnalogueGain": 2.0
-            }
-        )
-        self.picam2.configure(config)
-        self.picam2.start()
-
-        # Start unified processing thread
+        # Start frame capture thread
         self.running = True
-        self.process_thread = threading.Thread(target=self._process_loop)
-        self.process_thread.daemon = True
-        self.process_thread.start()
+        self.capture_thread = threading.Thread(target=self._capture_frames)
+        self.capture_thread.daemon = True
+        self.capture_thread.start()
 
-        time.sleep(1)
-        print("Unified camera initialized successfully")
+        # Wait a moment for first frame
+        time.sleep(2)
 
-    def _process_loop(self):
-        """Single thread handling both frame capture and QR detection"""
-        frame_count = 0
+        print("Camera initialization complete")
 
+    def _capture_frames(self):
+        """Continuously capture frames in a background thread using rpicam-jpeg"""
         while self.running:
             try:
-                # Capture both streams
-                request = self.picam2.capture_request()
+                # Use rpicam-jpeg to capture a frame with specified width and height
+                subprocess.run([
+                    "rpicam-jpeg",
+                    "--output", self.temp_file,
+                    "--nopreview",
+                    "--timeout", "1",
+                    "--width", str(self.width * 2),  # Capture at higher res for better QR reading
+                    "--height", str(self.height * 2),
+                    "--quality", "85"  # Moderate quality for source image
+                ], check=True, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
 
-                # Main frame for display
-                main_frame = request.make_array("main")
-                main_frame = cv2.cvtColor(main_frame, cv2.COLOR_RGB2BGR)
+                # Read the captured frame
+                frame = cv2.imread(self.temp_file)
 
-                # QR frame (grayscale, optimized)
-                lores_frame = request.make_array("lores")
-                qr_frame = cv2.cvtColor(lores_frame, cv2.COLOR_YUV420p2GRAY)
-
-                request.release()
-
-                # Update frames
-                with self.frame_lock:
-                    self.current_frame = main_frame
-                    self.qr_frame = qr_frame
-
-                # QR detection (every 3rd frame to balance performance)
-                frame_count += 1
-                if frame_count % 3 == 0 and self._should_scan_qr():
-                    self._detect_qr(qr_frame)
-
-                time.sleep(1/30)  # 30 FPS
-
+                if frame is not None:
+                    with self.frame_lock:
+                        self.current_frame = frame
+                    print("Frame captured successfully:", frame.shape)
+                else:
+                    print("Failed to read captured frame")
             except Exception as e:
-                print(f"Error in processing loop: {e}")
-                time.sleep(0.1)
+                print(f"Error capturing frame: {e}")
 
-    def _should_scan_qr(self):
-        """Check if we should scan for QR codes (respects cooldown)"""
-        if not self.qr_scanning_enabled:
-            return False
-
-        current_time = time.time()
-        return (current_time - self.last_qr_time) >= self.qr_cooldown
-
-    def _detect_qr(self, frame):
-        """Detect QR codes in the given frame"""
-        try:
-            # Multiple processing techniques for better detection
-            detection_frames = [
-                frame,  # Original
-                cv2.adaptiveThreshold(frame, 255, cv2.ADAPTIVE_THRESH_GAUSSIAN_C, cv2.THRESH_BINARY, 11, 2),
-                cv2.GaussianBlur(frame, (3, 3), 0)
-            ]
-
-            for processed_frame in detection_frames:
-                qr_codes = decode(processed_frame)
-                if qr_codes:
-                    for qr in qr_codes:
-                        qr_data = qr.data.decode('utf-8')
-                        self._handle_qr_detection(qr_data)
-                        return  # Stop after first successful detection
- 
-        except Exception as e:
-            print(f"QR detection error: {e}")
-
-    def _handle_qr_detection(self, qr_data):
-        """Handle a successful QR code detection"""
-        current_time = time.time()
-
-        # Avoid duplicate detections
-        if qr_data == self.last_qr_data and (current_time - self.last_qr_time) < 1.0:
-            return
-
-        print(f"QR Code detected: {qr_data}")
-
-        # Update state
-        self.last_qr_data = qr_data
-        self.last_qr_time = current_time
-
-        # Add to queue (thread-safe)
-        try:
-            self.qr_data_queue.put_nowait({
-                'data': qr_data,
-                'timestamp': current_time
-            })
-        except queue.Full:
-            # Remove oldest item and add new one
-            try:
-                self.qr_data_queue.get_nowait()
-                self.qr_data_queue.put_nowait({
-                    'data': qr_data,
-                    'timestamp': current_time
-                })
-            except queue.Empty:
-                pass
+            time.sleep(0.017)  # Capture at ~10 FPS to avoid overloading the system
 
     def get_frame(self):
-        """Get frame for web streaming with status overlay"""
+        """Get the latest frame from the camera, resized and compressed"""
         with self.frame_lock:
             if self.current_frame is None:
+                # Return a blank frame if no frame is available
                 blank = np.zeros((self.height, self.width, 3), dtype=np.uint8)
-                cv2.putText(blank, "No Camera Feed", (self.width//4, self.height//2),
+                # Add text to indicate no camera feed
+                cv2.putText(blank, "No Camera Feed", (int(self.width/4), int(self.height/2)),
                            cv2.FONT_HERSHEY_SIMPLEX, 0.5, (255, 255, 255), 1)
                 _, buffer = cv2.imencode('.jpg', blank, [cv2.IMWRITE_JPEG_QUALITY, self.jpeg_quality])
                 return buffer.tobytes()
 
+            # Return a resized copy of the current frame
             frame = self.current_frame.copy()
 
-        # Resize for display
+        # Resize the frame to the specified dimensions
         frame = cv2.resize(frame, (self.width, self.height))
 
-        # Add status overlay
-        self._add_status_overlay(frame)
-
-        # Encode
+        # Encode the frame with specified quality
         _, buffer = cv2.imencode('.jpg', frame, [cv2.IMWRITE_JPEG_QUALITY, self.jpeg_quality])
         return buffer.tobytes()
 
-    def _add_status_overlay(self, frame):
-        """Add QR scanning status to frame"""
-        current_time = time.time()
-
-        if not self.qr_scanning_enabled:
-            status = "QR Scanning: Disabled"
-            color = (0, 0, 255)  # Red
-        elif (current_time - self.last_qr_time) < self.qr_cooldown:
-            remaining = self.qr_cooldown - (current_time - self.last_qr_time)
-            status = f"QR Cooldown: {remaining:.1f}s"
-            color = (0, 165, 255)  # Orange
-        else:
-            status = "QR Scanning: Active"
-            color = (0, 255, 0)  # Green
-
-        cv2.putText(frame, status, (10, 30), cv2.FONT_HERSHEY_SIMPLEX, 0.7, color, 2)
-
-        # Show last detected QR
-        if self.last_qr_data:
-            qr_text = f"Last QR: {self.last_qr_data[:20]}..."
-            cv2.putText(frame, qr_text, (10, 60), cv2.FONT_HERSHEY_SIMPLEX, 0.5, (255, 255, 255), 1)
-
-    def get_latest_qr(self):
-        """Get the most recent QR detection (non-blocking)"""
-        try:
-            return self.qr_data_queue.get_nowait()
-        except queue.Empty:
-            return None
-
-    def has_qr_data(self):
-        """Check if there's QR data available"""
-        return not self.qr_data_queue.empty()
-
-    def clear_qr_queue(self):
-        """Clear all pending QR detections"""
-        while not self.qr_data_queue.empty():
-            try:
-                self.qr_data_queue.get_nowait()
-            except queue.Empty:
-                break
-
-    def set_qr_scanning(self, enabled):
-        """Enable/disable QR scanning"""
-        self.qr_scanning_enabled = enabled
-        if not enabled:
-            self.clear_qr_queue()
-        print(f"QR scanning {'enabled' if enabled else 'disabled'}")
-
-    def set_qr_cooldown(self, seconds):
-        """Set QR detection cooldown period"""
-        self.qr_cooldown = seconds
-        print(f"QR cooldown set to {seconds} seconds")
-
     def get_raw_frame(self):
-        """Get raw frame for other processing"""
+        """Get the raw frame for processing (not encoded)"""
         with self.frame_lock:
-            return self.current_frame.copy() if self.current_frame is not None else None
+            if self.current_frame is None:
+                return None
+            return self.current_frame.copy()
 
     def release(self):
-        """Clean up resources"""
-        print("Releasing unified camera resources...")
+        """Release the camera resources"""
+        print("Releasing camera resources...")
         self.running = False
+        if self.capture_thread.is_alive():
+            self.capture_thread.join(timeout=1.0)
 
-        if self.process_thread.is_alive():
-            self.process_thread.join(timeout=2.0)
-
+        # Clean up temporary directory
         try:
-            self.picam2.stop()
-            self.picam2.close()
+            import shutil
+            shutil.rmtree(self.temp_dir)
         except Exception as e:
-            print(f"Error closing camera: {e}")
-
-        self.clear_qr_queue()
-        print("Unified camera released")
+            print(f"Error cleaning temporary directory: {e}")
