+++ requirements.txt
@@ -1,5 +1,6 @@
 flask_mysqldb
 pygame
+picamera2
 opencv-python==4.8.0.76
 pyzbar==0.1.9
 numpy==1.24.3
